fix(chart-legend): fix according to the Pull Request Review
