+++ packages/charts/src/components/legend/index.ts
@@ -1,1 +1,2 @@
 export * from './legend'
+export * from './legend-item'
