+++ packages/charts/src/components/legend/legend.tsx
@@ -5,15 +5,14 @@ import {
   useCallback,
   useImperativeHandle,
   useMemo,
-  useRef,
   useState,
 } from 'react'
 import { defaultColorPreset } from '../../theme/colors'
 import type { EChartsOption } from 'echarts'
 import '../../theme/components/legend.css'
-import { checkAllSelected, getHoverColor } from '../../utils/legend'
-import { IconCheckSmall } from '@vtex/shoreline'
+import { checkAllSelected } from '../../utils/legend'
 import type ReactECharts from 'echarts-for-react'
+import { LegendItem } from './legend-item'
 
 export const Legend = forwardRef<LegendHandle, LegendProps>(
   function Legend(props, ref) {
@@ -28,25 +27,20 @@ export const Legend = forwardRef<LegendHandle, LegendProps>(
         initialState.push({ serie: String(serie.name), state: undefined })
     })
 
-    const [seriesState, setSeriesState] =
-      useState<LegendStateType>(initialState)
+    const [seriesState, setSeriesState] = useState<LegendState>(initialState)
 
     /**
      * Makes an array of colors to track the colors of each serie
      */
     const colors = useMemo(() => {
-      const colorsOut: string[] = []
       let index = 0
-      series.forEach((serie) => {
+      return series.map((serie) => {
         if (!serie.color) {
-          colorsOut.push(defaultColorPreset[index])
-          index++
-        } else {
-          if (isString(serie.color)) colorsOut.push(serie.color)
-          else colorsOut.push(serie.color[0])
+          return defaultColorPreset[index++]
         }
+        if (isString(serie.color)) return serie.color
+        return serie.color[0]
       })
-      return colorsOut
     }, [series, seriesState])
 
     /**
@@ -59,11 +53,11 @@ export const Legend = forwardRef<LegendHandle, LegendProps>(
      *
      * @param {number} index - Index of the serie to change the state.
      * @param {string} type - Type of action to perform. Can be 'toggle', 'selectAll' or 'exclusive'.
-     * @returns {LegendStateType} The new state of the legend.
+     * @returns {LegendState} The new state of the legend.
      */
     const changeState = useCallback(
-      (index: number, type: LegendActionType['type']) => {
-        const newState = [...seriesState] as LegendStateType
+      (index: number, type: LegendAction['type']) => {
+        const newState = [...seriesState] as LegendState
         if (type === 'selectAll') {
           newState.forEach((serie) => {
             serie.state = undefined
@@ -89,7 +83,7 @@ export const Legend = forwardRef<LegendHandle, LegendProps>(
       ref,
       () => ({
         setState: (index: number, action: string) => {
-          changeState(index, action as LegendActionType['type'])
+          changeState(index, action as LegendAction['type'])
         },
       }),
       [changeState]
@@ -104,16 +98,15 @@ export const Legend = forwardRef<LegendHandle, LegendProps>(
 
         const on: string[] = []
         const off: string[] = []
-        const un: string[] = []
+        const checked: string[] = []
         newState.forEach((serie) => {
-          console.log(serie.serie, serie.state)
           if (serie.state === true) on.push(serie.serie)
-          if (serie.state === false) off.push(serie.serie)
-          if (serie.state === undefined) un.push(serie.serie)
+          else if (serie.state === false) off.push(serie.serie)
+          else checked.push(serie.serie)
         })
 
         let action = ''
-        if (un.length !== 0) {
+        if (checked.length !== 0) {
           action = 'exclusive'
         } else if (on.length === 1 && on[0] === name) {
           action = 'selectAll'
@@ -121,10 +114,7 @@ export const Legend = forwardRef<LegendHandle, LegendProps>(
           action = 'toggle'
         }
 
-        const settedState = changeState(
-          index,
-          action as LegendActionType['type']
-        )
+        const settedState = changeState(index, action as LegendAction['type'])
 
         chart.dispatchAction({
           type: 'legendToggleSelect',
@@ -147,9 +137,7 @@ export const Legend = forwardRef<LegendHandle, LegendProps>(
               key={index}
               serie={serie.serie}
               onClick={onClick}
-              selected={
-                seriesState[index] ? seriesState[index].state : undefined
-              }
+              selected={seriesState[index].state}
               color={colors[index]}
               index={index}
             />
@@ -160,15 +148,15 @@ export const Legend = forwardRef<LegendHandle, LegendProps>(
   }
 )
 
-export type LegendStateType =
+export type LegendState =
   | { serie: string; state: boolean }[]
   | { serie: string; state: undefined }[]
 
-export type LegendActionType = {
+export type LegendAction = {
   index: number
   name: string
   type: 'toggle' | 'selectAll' | 'exclusive'
-  state: LegendStateType
+  state: LegendState
   chartId: string
 }
 
@@ -176,71 +164,10 @@ export type LegendHandle = {
   setState: (index: number, action: string) => void
 }
 
-export type LegendOptions = {
+type LegendOptions = {
   series: EChartsOption['series']
   chartRef: React.RefObject<ReactECharts>
 }
 
-export type LegendProps = LegendOptions &
-  Omit<ComponentPropsWithoutRef<'div'>, 'onClick'>
-
-function LegendItem({
-  serie,
-  onClick,
-  selected,
-  color,
-  index,
-  ...otherProps
-}: LegendItemProps) {
-  const buttonRef = useRef<HTMLButtonElement>(null)
-  const [hover, setHover] = useState<boolean>(false)
-
-  const backgroundColor = useMemo(() => {
-    if (selected === false) return 'transparent'
-    if (!hover) return color
-    return getHoverColor(color)
-  }, [selected, hover, color])
-
-  const handleClick = useCallback(
-    (_e: React.MouseEvent) => {
-      onClick(String(serie))
-    },
-    [serie, onClick]
-  )
-
-  return (
-    <div
-      data-sl-chart-legend-container
-      onClick={handleClick}
-      onMouseDown={(e) => e.preventDefault()}
-      onMouseEnter={() => setHover(true)}
-      onMouseLeave={() => setHover(false)}
-      {...otherProps}
-    >
-      <button
-        ref={buttonRef}
-        data-sl-chart-legend-button
-        style={{
-          backgroundColor: backgroundColor,
-          border: selected !== false ? 'none' : 'var(--sl-border-base-strong)',
-        }}
-      >
-        {selected === true ? (
-          <IconCheckSmall data-sl-chart-legend-check />
-        ) : null}
-      </button>
-      <span data-sl-chart-legend-text>{serie}</span>
-    </div>
-  )
-}
-
-type LegendItemOptions = {
-  serie: string
-  onClick: (name: string) => void
-  selected?: boolean
-  color: string
-  index: number
-}
-
-type LegendItemProps = LegendItemOptions &
+type LegendProps = LegendOptions &
   Omit<ComponentPropsWithoutRef<'div'>, 'onClick'>
